Separated AuthError into different error files for readability, removed unnecessary comments and updated exports
