+++ lib/msal-core/src/index.ts
@@ -7,4 +7,11 @@ export { ErrorMessage } from "./ErrorMessage";
 export { TokenResponse } from "./RequestInfo";
 export { Authority } from "./Authority";
 export { CacheResult } from "./UserAgentApplication";
-export { AuthError, ClientAuthError, ConfigurationError, InteractionRequiredAuthError, ServerError, ClaimsRequiredAuthError } from "./AuthError";
+
+// Errors
+export { AuthError } from "./error/AuthError";
+export { ClientAuthError } from "./error/ClientAuthError";
+export { ServerError } from "./error/ServerError";
+export { ClientConfigurationError } from "./error/ClientConfigurationError";
+export { InteractionRequiredAuthError } from "./error/InteractionRequiredError";
+export { ClaimsRequiredAuthError } from "./error/ClaimsRequiredError";
